backend/pyproject.toml
[project]
name = "backend"
version = "0.0.1"
description = "Videogame best deals finder"
authors = [
    {name = "David 'TM' Bellini",email = "david@thinkingmonkey.eu"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

